MiG-15 version 3.0: add support for Bombable
This support is preliminary at this point, please report any bugs to
the SourceForge bug tracker or, failing that, to
flightgear-devel@lists.sourceforge.net.

The new major version number is because new properties are now being
transmitted over the multiplayer protocol.

It should theoretically be possible to engage in dogfights over
multiplayer now but I have not been able to test that yet; I have
however tested the bombable scenarios in single player mode and
confirmed that the MiG-15 can now deal and receive damage from
other aircraft.